requirements.txt
tensorflow-addons
tensorboard_plugin_profile
dtw
pandas
numpy
scipy
jupyter
matplotlib
seaborn
Pillow
scikit-learn
librosa
tqdm
pydub
soundfile